Move loading action into misc.js
