--- src/scripts/actions/general.js
@@ -1,8 +1,0 @@
-/**
- * @param {boolean} isLoading
- * @returns {{type: string, isLoading: boolean}}
- */
-export const setIsLoading = isLoading => ({
-  type: 'SET_IS_LOADING',
-  isLoading,
-});
